Set up  functions redering to DOM(test)
