+++ src/Functions.svelte
@@ -14,7 +14,7 @@ export function getMode( scale: string[], scaleDegree: number) {
   let notes = scale.slice(0,scaleDegree - 1)
   let flippedNotes = scale.slice(- scale.length + (scaleDegree - 1))
   mode = notes.map(note => {flippedNotes.push(note)})
-  console.log(flippedNotes)
+//   console.log(flippedNotes)
   return flippedNotes
   
 }
